refactor: 컨벤션에 맞춰 코드 리팩토링
