+++ src/services/apis/userApi.ts
@@ -1,38 +1,5 @@
 import { currentAuth } from '@/services/apis/authApi';
-
-interface IAddress {
-  address: string;
-  roadAddress: string;
-  latitude: number;
-  longitude: number;
-  sido: string;
-  bname: string;
-  sigungu: string;
-}
-
-interface IUserPlaceLog {
-  userName: string;
-  id: number;
-  name: string;
-  imageUrl: string;
-  address: IAddress;
-}
-
-interface IUserBookmarkPlace extends IUserPlaceLog {
-  category: 'TOUR';
-}
-
-export interface IUser {
-  name: string;
-  instagramId: string;
-  imageUrl: string;
-  description: string;
-  follower: number;
-  following: number;
-  userPlaceLogDtoList: IUserPlaceLog[];
-  userBookmarkPlaceLogDtoList: IUserPlaceLog[];
-  userBookmarkPlaceDtoList: IUserBookmarkPlace[];
-}
+import { IUpdateUser, IUser } from '@/services/apis/types/userAPI.type';
 
 class User {
   private http = currentAuth;
@@ -40,6 +7,16 @@ class User {
   /* 로그인된 사용자 정보 가져오기 */
   async getUser(): Promise<IUser> {
     const response = await this.http.get('/users');
+    return response.data;
+  }
+
+  async deleteUser() {
+    const response = await this.http.delete('/users');
+    return response.data;
+  }
+
+  async patchUser(userData: IUpdateUser): Promise<IUpdateUser> {
+    const response = await this.http.patch('/users', userData);
     return response.data;
   }
 }
