--- src/services/queryKeys/index.ts
@@ -1,1 +1,0 @@
-export { userKeys } from '@/services/queryKeys/userKeys';
